DataFeed/requirements.txt
pandas
numpy
searchtweets
unidecode
newsapi-python
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
unidecode
apscheduler
google-cloud-translate
pymysql
sqlalchemy